Fix march EBA buff not being applied to attack caused by the attack
We can't add it there directly, as if the master with someone that deals
extra dmg, it would be affected. While it should only effect attacks
after.

Aven has also seen a change in behavior due to his FUA not interuppting
attacks anymore. This causes changes in dmg, but is more correct I think